server/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pydantic_core
python-dotenv
mysql-connector-python
bcrypt
beautifulsoup4
httpx
requests
googlesearch-python
google
google-search